requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pdf2image==1.14.0
pillow==10.2.0
pytesseract==0.3.10
Werkzeug==2.3.8
